Create Question 1
All triangles have 3 sides and 3 angles which always add up to 180°. They are classified in two major 
ways that is:-

1) By the number of equal sides they have:
  • Scalene - no equal sides
  • Isosceles  - two equal sides
  • Equilateral - three (all) side equal

2) By the types of angles they have:

  • Acute triangle - all angles < 90
  • Right triangle - one angle ==90
  • Obtuse triangle - one angle > 90

You have been hired as a programmer by the Ministry of Education to develop a script in JavaScript for
Form 1 students.  The said program should test students’ knowledge on the types of triangles, 
classified with regards to  the number of equal sides they have. 

The program that you are to develop MUST prompt the user for a Triangle (inclusive of their 
dimensions), then print the type of triangle that conforms to (that is, Scalene, 
 Isosceles and Equilateral) and lastly print the number of times each triangle was entered. 

Note, for each triangle you must prompt the user to enter all of its dimensions i.e., side 1, side 2 
and side 3. Also, a valid triangle must have sides greater than 0. Thus, if the user enters any 
side that is not greater than 0 then the triangle would be invalid and your program must state 
such.